Hooks design pattern refactor
