+++ src/client/features/team/components/available-golfers-list/search-bar.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import { InputAdornment } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 import SearchIcon from '@mui/icons-material/Search';
-import { SearchBarWrapper, StyledIconButton, StyledSearchBar } from './styled';
+import { InputAdornment } from '@mui/material';
+import React from 'react';
+import { SearchBarWrapper, StyledIconButton, StyledSearchBar } from '../styled';
 
 interface Props {
   searchTerm: string;
